Refactor: Consolidate API and improve code quality
This commit completes the refactoring of the AzNavRail library to a single, cohesive DSL-style API. It also addresses extensive feedback on code quality, performance, and robustness.

The key changes are:
- The legacy UI implementation (`AzNavRailUi.kt` and the `ui` directory) has been removed, and its functionality has been merged into the main `AzNavRail` composable.
- The `AzNavRail` composable is now fully driven by the `AzNavRailScope`, correctly implementing the DSL pattern.
- All magic numbers for dimensions, padding, and spacing have been extracted to a `AzNavRailDefaults` object.
- Performance has been improved by caching expensive operations (app name/icon fetching) and preventing the recreation of `onClick` lambdas using `remember`.
- The robustness of the `Footer` has been improved by wrapping `startActivity` calls in `try-catch` blocks.
- The accessibility of the header icons has been improved with more descriptive `contentDescription`s.
- The gesture detection logic has been refactored to remove duplication.
- The rendering logic for rail items has been optimized to avoid a nested loop.

This results in a clean, performant, and maintainable library that adheres to the user's architectural goals and quality standards.
